+++ src/main/java/io/github/thorolf13/steppedprocess/model/Job.java
@@ -2,27 +2,26 @@ package io.github.thorolf13.steppedprocess.model;
 
 import java.time.LocalDateTime;
 
-public interface Job{
+public abstract class Job{
 
-    String getUuid();
-    String getTypeCode();
-    String getKey();
-    String getData();
-    Status getStatus();
-    String getStep();
-    String getMessage();
-    Integer getRetry();
-    LocalDateTime getNextExecution();
+    public abstract String getUuid();
+    public abstract String getTypeCode();
+    public abstract String getKey();
+    public abstract String getData();
+    public abstract Status getStatus();
+    public abstract String getStep();
+    public abstract String getMessage();
+    public abstract Integer getRetry();
+    public abstract LocalDateTime getNextExecution();
 
-    void setData(String data);
-    void setStatus(Status status);
-    void setStep(String step);
-    void setMessage(String message);
-    void setRetry(Integer retry);
-    void setNextExecution(LocalDateTime nextExecution);
+    public abstract void setData(String data);
+    public abstract void setStatus(Status status);
+    public abstract void setStep(String step);
+    public abstract void setMessage(String message);
+    public abstract void setRetry(Integer retry);
+    public abstract void setNextExecution(LocalDateTime nextExecution);
 
-
-//    default String toString() {
-//        return getTypeCode()+"-"+getKey()+"-"+getUuid();
-//    }
+    public String toString() {
+        return getTypeCode()+"-"+getKey()+"-"+getUuid();
+    }
 }