+++ src/main/java/io/github/thorolf13/steppedprocess/provided/JobRepository.java
@@ -7,13 +7,12 @@ import io.github.thorolf13.steppedprocess.model.Status;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 
-public interface JobRepository {
-    Optional<Job> getByUuid(String uuid);
-    List<Job> getByTypeCodeAndKey(String typeCode, String key);
+public interface JobRepository<I extends Job>{
+    Optional<I> findOneByUuid(String uuid);
+    List<I> findAllByTypeCodeAndKey(String typeCode, String key);
     int countByTypeCodeAndStatus(String type, Status status);
-    List<Job> getByTypeCodeAndStatus(String type, Status status);
-    Job save(Job job);
-    Job createJob(String typeCode, String key, String data, LocalDateTime executionTime);
+    List<I> findAllByTypeCodeAndStatus(String type, Status status);
+    I saveJob(I job);
+    I createJob(String typeCode, String key, String data, LocalDateTime executionTime);
 }